+++ lib/rubocop/cached_data.rb
@@ -38,7 +38,7 @@ module RuboCop
     def message(offense)
       # JSON.dump will fail if the offense message contains text which is not
       # valid UTF-8
-      offense.message.scrub
+      offense.message.dup.force_encoding(::Encoding::UTF_8).scrub
     end
 
     # Restore an offense object loaded from a JSON file.